requirements.txt
discord.py>=2.3.0
aiohttp>=3.9.0
python-dateutil>=2.8.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.100.0
todoist-api-python>=2.1.0
boto3>=1.34.0
